create updateDB script per la produzione

+++ lib/tasks/initDB.rake
@@ -5,12 +5,12 @@ namespace :initDB do
 
   # Script che svuota il db e lo ri-popola con una struttura vergine
   task :init_db => :environment do
-    # if Rails.env != 'test'
-    #   puts 'Are you sure you want to re-initialize the database? [y/n]'
-    #   response = STDIN.gets.chomp
-    # end
+    if Rails.env != 'test'
+      puts 'Are you sure you want to re-initialize the database? [y/n]'
+      response = STDIN.gets.chomp
+    end
 
-    # if Rails.env.test? || response.downcase == 'y'
+    if Rails.env.test? || response.downcase == 'y'
       # Svuoto il db
       MetalValue.delete_all
       Deadline.delete_all
@@ -22,12 +22,9 @@ namespace :initDB do
       # Salvo il valore corrente dell'oro e dell'argento
       GetRealTimeGoldPriceCommand.call
 
-      first_user = User.create!(email: 'm.battistelli@aigrading.com', password: '111111', first_name: 'Michele', last_name: 'Battistelli', role: 'admin')
 
       # Creo la prima organizzazione
       first_organization = Organization.where(name: 'Famiglia Battistelli').first_or_create!
-
-      Membership.create!(user_id: first_user.id, organization_id: first_organization.id, role: 'editor')
 
 
       # Creo le prime voci di spesa della prima organizzazione
@@ -188,9 +185,9 @@ namespace :initDB do
         initial_amount: 4987.48,
         ordering_number: 14
       ).first_or_create!
-    # else
-    #   puts '****** RAKE ABORTED!!! ******'
-    # end
+    else
+      puts '****** RAKE ABORTED!!! ******'
+    end
   end
 
 
@@ -213,7 +210,7 @@ namespace :initDB do
     file_paths = file_paths.filter{ | file | file.include?('xlsx') && !file.to_s.downcase.include?('andamento') }
 
     expense_items_by_colors = ExpenseItem.all.map{ | expense_item | [ expense_item.color[1..-1], expense_item.id ] }.to_h
-    ImportCountMovementsFromXlsxFileCommand.call(Count.find(100), file_paths, expense_items_by_colors)
+    ImportCountMovementsFromXlsxFileCommand.call(Count.find(1), file_paths, expense_items_by_colors)
   end
